Updated database types

+++ src/sendable/data/Position.java
@@ -19,15 +19,15 @@ public class Position implements Sendable {
     private int tableID;
     private int uid;
     private long time;
-    private int xPos;
-    private int yPos;
-    private int zPos;
+    private double xPos;
+    private double yPos;
+    private double zPos;
 
     public Position() {
         this(0, 0, 0, 0, 0);
     }
 
-    public Position(int uid, long time, int xPos, int yPos, int zPos) {
+    public Position(int uid, long time, double xPos, double yPos, double zPos) {
         this.uid = uid;
         this.time = time;
         this.xPos = xPos;
@@ -37,7 +37,7 @@ public class Position implements Sendable {
     }
 
     @Column(name = "X_POS", nullable = false)
-    public int getxPos() {
+    public double getxPos() {
         return xPos;
     }
 
@@ -46,7 +46,7 @@ public class Position implements Sendable {
     }
 
     @Column(name = "Y_POS", nullable = false)
-    public int getyPos() {
+    public double getyPos() {
         return yPos;
     }
 
@@ -55,7 +55,7 @@ public class Position implements Sendable {
     }
 
     @Column(name = "Z_POS", nullable = false)
-    public int getzPos() {
+    public double getzPos() {
         return zPos;
     }
 
@@ -125,9 +125,9 @@ public class Position implements Sendable {
         int result = 11;
         result = 37 * result + uid;
         result = 37 * result + Long.valueOf(time).hashCode();
-        result = 37 * result + xPos;
-        result = 37 * result + yPos;
-        result = 37 * result + zPos;
+        result = (int) (37 * result + xPos);
+        result = (int) (37 * result + yPos);
+        result = (int) (37 * result + zPos);
         return result;
     }
 }
